Changing output from CSV to PEM
We only require the certificate and no other metadata about it, so instead of
writing a CSV with the certificate as Base64 encoded DER ASN.1, we just write
each out individually.

This may change in future if the lack of file structure hits filesystem limits
as all certificates are placed into a separate directory. We could then batch
like the CSV file originally did to claim back further storage of certificates
before reaching any limits again.